requirements.txt
pygame==2.6.1
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
Pillow==10.2.0
